requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
MouseInfo==0.1.3
Pillow==8.4.0
PyAudio==0.2.11
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python-decouple==3.5
python-dotenv==0.19.2
python3-xlib==0.15
pyttsx3==2.71
pytweening==1.0.4
pywhatkit==5.2
requests==2.26.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
urllib3==1.26.7
wikipedia==1.4.0
